[D83T-29] 키워드칩 Composable 앱 전역에서 사용할 수 있도록 변경
